requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
cerebras-cloud-sdk
ultralytics
python-dotenv
